feat: merge connect and sign-in flow
